记录员工企业变更明细
git-svn-id: svn://120.55.85.221/doooly/trunk/APP2.0/doooly@18886 97f0348b-149b-491d-833b-dba32c3ac2c0

+++ src/main/java/com/doooly/business/activity/impl/XingFuJiaoHangActivityService.java
@@ -83,7 +83,7 @@ public class XingFuJiaoHangActivityService extends AbstractActivityService {
 					// 员工类型
 					user.setType((short) AdUser.TYPE_EMPLOYEE);
 					userService.syncUserASystem(user);
-					if (user.getOldGroupNum() != null && user.getGroupNum() != user.getOldGroupNum()) {
+					if (user.getOldGroupNum() != null && user.getGroupNum().longValue() != user.getOldGroupNum().longValue()) {
 						enterpriseChangeDao.insert(new AdUserEnterpriseChange(user.getId(), user.getOldGroupNum(),
 								user.getGroupNum(), Integer.valueOf(user.getIsActive())));
 					}
